Create a new Merge class for QScript merging functions.
